Update codex DEV, and add new entries. Date test.
I think warnings should be parsed into different ones. @god-hephaestus wdyt?

Entry dates can be to the future which I like, they are only for us to sort anyway, and we can backfill later.

We should update the dates when they are actually updated though.

We shouldn't have entries that conflict as well, mostly because IDK how it impacts the filters.